API

List of MendelKinship.jl's functions.

Index

MendelKinship.Kinship

Functions

MendelKinship.Kinship — Function

This is the wrapper function for the Kinship analysis option. Keyword options should be specified in the control file.

Missing docstring.

Missing docstring for compare_kinships. Check Documenter's build log for details.

Missing docstring.

Missing docstring for theoretical_coefficients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for kinship_matrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for delta7_matrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jacquard_coefficients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cotwin_extension!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for identity_state. Check Documenter's build log for details.

Missing docstring.

Missing docstring for compute_fishers_z. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plot_fisher_z. Check Documenter's build log for details.